modified:   app/Http/Controllers/AdminController.php 	modified:   app/Http/Controllers/CustomerAuthController.php 	modified:   app/Http/Controllers/CustomerController.php 	modified:   app/Http/Controllers/TechnicianAuthController.php 	modified:   app/Http/Controllers/TechnicianController.php 	modified:   app/Models/Technician.php 	renamed:    database/migrations/2024_08_12_133710_create_customers_table.php -> database/migrations/2024_10_16_175830_create_customers_table.php 	renamed:    database/migrations/2024_08_13_111607_create_technicians_table.php -> database/migrations/2024_10_16_175947_create_technicians_table.php 	modified:   public/css/Admin/6 - ViewProfile.css 	modified:   public/css/Customer/7 - Repairlist.css 	modified:   resources/views/Admin/2 - UserManagement.blade.php 	modified:   resources/views/Admin/6 - ViewProfile.blade.php 	new file:   resources/views/Admin/7 - ViewProfile.blade.php 	modified:   resources/views/Customer/7 - Repairlist.blade.php 	modified:   resources/views/Technician/6 - ManageProfile.blade.php 	modified:   routes/web.php

+++ app/Http/Controllers/TechnicianController.php
@@ -678,29 +678,31 @@ class TechnicianController extends Controller
             $existingServices = RepairShop_Services::where('technician_id', $technician->id)->get();
             
             // Iterate over the provided services
-            foreach ($request->input('service') as $index => $service) {
-                if (!empty($service)) {
-                    // Check if the service already exists (based on ID or any other unique attribute)
-                    $existingService = $existingServices->get($index);
+            if($request->input('service')){
+                foreach ($request->input('service') as $index => $service) {
+                    if (!empty($service)) {
+                        // Check if the service already exists (based on ID or any other unique attribute)
+                        $existingService = $existingServices->get($index);
 
-                    // Update the existing service or create a new one
-                    RepairShop_Services::updateOrCreate(
-                        [
-                            'id' => $existingService->id ?? null, // Use the existing ID if available
-                            'technician_id' => $technician->id,
-                        ],
-                        [
-                            'service' => $service,
-                        ]
-                    );
+                        // Update the existing service or create a new one
+                        RepairShop_Services::updateOrCreate(
+                            [
+                                'id' => $existingService->id ?? null, // Use the existing ID if available
+                                'technician_id' => $technician->id,
+                            ],
+                            [
+                                'service' => $service,
+                            ]
+                        );
+                    }
                 }
-            }
 
-            // If there are fewer services in the request than in the database, remove the extra ones
-            if ($existingServices->count() > count($request->input('service'))) {
-                $excessServices = $existingServices->skip(count($request->input('service')));
-                foreach ($excessServices as $excessService) {
-                    $excessService->delete();
+                // If there are fewer services in the request than in the database, remove the extra ones
+                if ($existingServices->count() > count($request->input('service'))) {
+                    $excessServices = $existingServices->skip(count($request->input('service')));
+                    foreach ($excessServices as $excessService) {
+                        $excessService->delete();
+                    }
                 }
             }
